Consolidate server knobs header files
